Refactor chat functionality with hooks for sending and receiving messages, loading state, and conversation history
