+++ avengers/frontend/webapp/src/app/assistants/featurecraft/components/newMessageBlock.js
@@ -21,13 +21,11 @@ export default function NewMessageBlock({onSendMessage, onReceiveMessage, conver
             try {
                 onSendMessage(newMessage);
                 setMessage("");
-                
                 // Reset the height of the textarea
                 const textarea = document.querySelector('textarea');
                 if (textarea) {
                     textarea.style.height = 'auto';
                 }
-                
                 const response = await axios.post("http://localhost:8080/chat", newMessage);
                 onReceiveMessage(response);
                 
@@ -60,5 +58,5 @@ export default function NewMessageBlock({onSendMessage, onReceiveMessage, conver
                 </div>
             </div>
         </div>
-    )
+    );
 }
