+++ avengers/frontend/webapp/src/app/assistants/featurecraft/hooks/useAssistChatSend.js
@@ -1,0 +1,14 @@
+import axios from "axios";
+
+export default function useAssistChatSend(messages, setMessages) {
+    const handleSendMessage = async (sentMessage) => {
+        setMessages((prevMessages) => [
+            ...prevMessages,
+            sentMessage.newMessage
+        ]);
+    };
+
+    return {
+        handleSendMessage
+    };
+}
